options/glibc: fix elf_greg_t definition

+++ options/glibc/include/sys/procfs.h
@@ -9,7 +9,7 @@
 extern "C" {
 #endif
 
-typedef unsigned long long elf_greg_t;
+typedef unsigned long elf_greg_t;
 
 #define ELF_NGREG (sizeof (struct user_regs_struct) / sizeof (elf_greg_t))
 typedef elf_greg_t elf_gregset_t[ELF_NGREG];
